bugfix: coverage of filled-in coordinates
The rows added to compensate for deletions in the mutated references were getting lost due to filtering on `mutated` which is `NA` for coordinates added after the shoft correction.  This was causing the coverage line to jump across the deletion instead if dropping to 0 and staying there throughout the delete region.